blockactive/tests/address-validator.test.ts
import {
  assert,
  describe,
  test,
  clearStore,
  beforeAll,
  afterAll
} from "matchstick-as/assembly/index"
import { Address } from "@graphprotocol/graph-ts"
import { AddACorruptAddress } from "../generated/schema"
import { AddACorruptAddress as AddACorruptAddressEvent } from "../generated/AddressValidator/AddressValidator"
import { handleAddACorruptAddress } from "../src/address-validator"
import { createAddACorruptAddressEvent } from "./address-validator-utils"

// Tests structure (matchstick-as >=0.5.0)
// https://thegraph.com/docs/en/developer/matchstick/#tests-structure-0-5-0

describe("Describe entity assertions", () => {
  beforeAll(() => {
    let _address = Address.fromString(
      "0x0000000000000000000000000000000000000001"
    )
    let newAddACorruptAddressEvent = createAddACorruptAddressEvent(_address)
    handleAddACorruptAddress(newAddACorruptAddressEvent)
  })

  afterAll(() => {
    clearStore()
  })

  // For more test scenarios, see:
  // https://thegraph.com/docs/en/developer/matchstick/#write-a-unit-test

  test("AddACorruptAddress created and stored", () => {
    assert.entityCount("AddACorruptAddress", 1)

    // 0xa16081f360e3847006db660bae1c6d1b2e17ec2a is the default address used in newMockEvent() function
    assert.fieldEquals(
      "AddACorruptAddress",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "_address",
      "0x0000000000000000000000000000000000000001"
    )

    // More assert options:
    // https://thegraph.com/docs/en/developer/matchstick/#asserts
  })
})
